+++ src/main/java/com/example/petbuddybackend/service/session/chat/ChatSessionTracker.java
@@ -1,4 +1,4 @@
-package com.example.petbuddybackend.service.chat.session;
+package com.example.petbuddybackend.service.session.chat;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.annotation.Scope;
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Component;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Set;
 
 @Component
 @Profile("!test")
@@ -17,24 +16,16 @@ public class ChatSessionTracker {
 
     private final Map<String, Long> subsToChatId = new HashMap<>();
 
-    public Set<String> getSubscriptionIds() {
-        return subsToChatId.keySet();
-    }
-
     public Map<String, Long> getSubscriptions() {
         return Collections.unmodifiableMap(subsToChatId);
     }
 
-    public void addSubscription(String sessionId, Long chatId) {
-        subsToChatId.put(sessionId, chatId);
+    public void addSubscription(String subscriptionId, Long chatId) {
+        subsToChatId.put(subscriptionId, chatId);
     }
 
-    public Long removeSubscription(String sessionId) {
-        return subsToChatId.remove(sessionId);
-    }
-
-    public Long getChatId(String sessionId) {
-        return subsToChatId.get(sessionId);
+    public Long removeSubscription(String subscriptionId) {
+        return subsToChatId.remove(subscriptionId);
     }
 
     public void clear() {
